+++ app/Models/Category.php
@@ -1,6 +1,4 @@
-<?php
-
-namespace App;
+<?php namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Builder;
@@ -15,21 +13,21 @@ use Illuminate\Database\Eloquent\Builder;
  * @property integer $parent_id
  * @property \Carbon\Carbon $created_at
  * @property \Carbon\Carbon $updated_at
- * @property-read \Illuminate\Database\Eloquent\Collection|\App\Article[] $articles
- * @method static \Illuminate\Database\Query\Builder|\App\Category whereId($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Category whereTitle($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Category whereSlug($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Category whereDescription($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Category whereParentId($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Category whereCreatedAt($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Category whereUpdatedAt($value)
+ * @property-read \Illuminate\Database\Eloquent\Collection|\App\Models\Article[] $articles
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Category whereId($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Category whereTitle($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Category whereSlug($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Category whereDescription($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Category whereParentId($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Category whereCreatedAt($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Category whereUpdatedAt($value)
  * @mixin \Eloquent
- * @property-read \App\File $logo
+ * @property-read \App\Models\File $logo
  * @property-read mixed $logo_url
  * @property integer $logo_id
- * @method static \Illuminate\Database\Query\Builder|\App\Category whereLogoId($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Category whereLogoId($value)
  * @property string $type
- * @method static \Illuminate\Database\Query\Builder|\App\Category whereType($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Category whereType($value)
  */
 class Category extends Model
 {
