+++ app/Models/Comment.php
@@ -1,6 +1,4 @@
-<?php
-
-namespace App\Models;
+<?php namespace App\Models;
 
 use Illuminate\Database\Eloquent\Builder;
 use Prettus\Repository\Contracts\Transformable;
