+++ app/Models/Permission.php
@@ -1,0 +1,28 @@
+<?php namespace App\Models;
+/**
+ * Created by PhpStorm.
+ * 
+ * User: shellus
+ * Date: 2016-03-09
+ * Time: 15:06
+ *
+ * @property integer $id
+ * @property string $name
+ * @property string $display_name
+ * @property string $description
+ * @property \Carbon\Carbon $created_at
+ * @property \Carbon\Carbon $updated_at
+ * @property-read \Illuminate\Database\Eloquent\Collection|\App\Models\Role[] $roles
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Permission whereId($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Permission whereName($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Permission whereDisplayName($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Permission whereDescription($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Permission whereCreatedAt($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Permission whereUpdatedAt($value)
+ * @mixin \Eloquent
+ */
+
+use Zizaco\Entrust\EntrustPermission;
+class Permission extends EntrustPermission
+{
+}
