+++ app/Models/Role.php
@@ -1,0 +1,28 @@
+<?php namespace App\Models;
+/**
+ * Created by PhpStorm.
+ * 
+ * User: shellus
+ * Date: 2016-03-09
+ * Time: 15:05
+ *
+ * @property integer $id
+ * @property string $name
+ * @property string $display_name
+ * @property string $description
+ * @property \Carbon\Carbon $created_at
+ * @property \Carbon\Carbon $updated_at
+ * @property-read \Illuminate\Database\Eloquent\Collection|\App\Models\User[] $users
+ * @property-read \Illuminate\Database\Eloquent\Collection|\App\Models\Permission[] $perms
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Role whereId($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Role whereName($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Role whereDisplayName($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Role whereDescription($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Role whereCreatedAt($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\Role whereUpdatedAt($value)
+ * @mixin \Eloquent
+ */
+use Zizaco\Entrust\EntrustRole;
+class Role extends EntrustRole
+{
+}
