+++ app/Models/User.php
@@ -1,6 +1,4 @@
-<?php
-
-namespace App;
+<?php namespace App\Models;
 
 use App\Notifications\ResetPasswordNotification;
 use Illuminate\Notifications\Notifiable;
@@ -18,23 +16,23 @@ use Zizaco\Entrust\Traits\EntrustUserTrait;
  * @property \Carbon\Carbon $updated_at
  * @property-read \Illuminate\Notifications\DatabaseNotificationCollection|\Illuminate\Notifications\DatabaseNotification[] $notifications
  * @property-read \Illuminate\Notifications\DatabaseNotificationCollection|\Illuminate\Notifications\DatabaseNotification[] $unreadNotifications
- * @method static \Illuminate\Database\Query\Builder|\App\User whereId($value)
- * @method static \Illuminate\Database\Query\Builder|\App\User whereName($value)
- * @method static \Illuminate\Database\Query\Builder|\App\User whereEmail($value)
- * @method static \Illuminate\Database\Query\Builder|\App\User wherePassword($value)
- * @method static \Illuminate\Database\Query\Builder|\App\User whereRememberToken($value)
- * @method static \Illuminate\Database\Query\Builder|\App\User whereCreatedAt($value)
- * @method static \Illuminate\Database\Query\Builder|\App\User whereUpdatedAt($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\User whereId($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\User whereName($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\User whereEmail($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\User wherePassword($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\User whereRememberToken($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\User whereCreatedAt($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\User whereUpdatedAt($value)
  * @mixin \Eloquent
  * @property string $api_token
  * @property string $register_ip
- * @property-read \Illuminate\Database\Eloquent\Collection|\App\Role[] $roles
- * @method static \Illuminate\Database\Query\Builder|\App\User whereApiToken($value)
- * @method static \Illuminate\Database\Query\Builder|\App\User whereRegisterIp($value)
+ * @property-read \Illuminate\Database\Eloquent\Collection|\App\Models\Role[] $roles
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\User whereApiToken($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\User whereRegisterIp($value)
  * @property integer $avatar_id
- * @property-read \App\File $avatar
+ * @property-read \App\Models\File $avatar
  * @property-read mixed $avatar_url
- * @method static \Illuminate\Database\Query\Builder|\App\User whereAvatarId($value)
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\User whereAvatarId($value)
  * @property-read mixed $role_name
  */
 class User extends Authenticatable
@@ -69,7 +67,7 @@ class User extends Authenticatable
      * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
      */
     public function avatar(){
-        return $this->belongsTo('App\File');
+        return $this->belongsTo('App\Models\File');
     }
 
     public function getAvatarUrlAttribute(){
@@ -87,7 +85,7 @@ class User extends Authenticatable
 
     public function roles()
     {
-        return $this->belongsToMany('App\Role');
+        return $this->belongsToMany('App\Models\Role');
     }
 
     /**
