--- app/Permission.php
@@ -1,28 +1,0 @@
-<?php namespace App;
-/**
- * Created by PhpStorm.
- * 
- * User: shellus
- * Date: 2016-03-09
- * Time: 15:06
- *
- * @property integer $id
- * @property string $name
- * @property string $display_name
- * @property string $description
- * @property \Carbon\Carbon $created_at
- * @property \Carbon\Carbon $updated_at
- * @property-read \Illuminate\Database\Eloquent\Collection|\App\Role[] $roles
- * @method static \Illuminate\Database\Query\Builder|\App\Permission whereId($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Permission whereName($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Permission whereDisplayName($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Permission whereDescription($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Permission whereCreatedAt($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Permission whereUpdatedAt($value)
- * @mixin \Eloquent
- */
-
-use Zizaco\Entrust\EntrustPermission;
-class Permission extends EntrustPermission
-{
-}
