--- app/Role.php
@@ -1,28 +1,0 @@
-<?php namespace App;
-/**
- * Created by PhpStorm.
- * 
- * User: shellus
- * Date: 2016-03-09
- * Time: 15:05
- *
- * @property integer $id
- * @property string $name
- * @property string $display_name
- * @property string $description
- * @property \Carbon\Carbon $created_at
- * @property \Carbon\Carbon $updated_at
- * @property-read \Illuminate\Database\Eloquent\Collection|\App\User[] $users
- * @property-read \Illuminate\Database\Eloquent\Collection|\App\Permission[] $perms
- * @method static \Illuminate\Database\Query\Builder|\App\Role whereId($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Role whereName($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Role whereDisplayName($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Role whereDescription($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Role whereCreatedAt($value)
- * @method static \Illuminate\Database\Query\Builder|\App\Role whereUpdatedAt($value)
- * @mixin \Eloquent
- */
-use Zizaco\Entrust\EntrustRole;
-class Role extends EntrustRole
-{
-}
